Documentation and code cleanup
