+++ rome.json
@@ -1,5 +1,8 @@
 {
   "$schema": "./node_modules/rome/configuration_schema.json",
+  "files": {
+    "ignore": ["packages/**/dist"]
+  },
   "formatter": {
     "enabled": true,
     "indentStyle": "space",
